llm_ad/requirements.txt
flask
pandas
langchain-community
langchain-openai
python-dotenv
numpy